worker: parallellize fetching revisions
